Re-organize -- move common functions to GA package
