+++ GAlgorithm/__init__.py
@@ -1,7 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from .population import AbstractIndividual
-from .population import Population
+from .population import (
+    AbstractIndividual,
+    Population,
+    initialize_population)
+
 from .geneticalgorithm import GeneticAlgorithm
 
 from .saver import load, dump
